requirements.txt
beautifulsoup4==4.13.4
openpyxl==3.1.5
python-dotenv==1.1.0
Requests==2.32.4
selenium==4.33.0
undetected_chromedriver==3.5.5
